requirements.txt
streamlit>=1.18.0
pandas>=1.3.0
plotly>=5.5.0
google-api-python-client>=2.40.0
oauth2client>=4.1.3
# Added Instagram scraping and analysis libraries
instaloader>=4.9.5
python-dateutil>=2.8.2
pillow>=9.0.0
wordcloud>=1.8.2.2
matplotlib>=3.5.0
# Keep these if you ever want to use OAuth in the future
# google-auth>=2.6.0
# google-auth-oauthlib>=0.4.6
# google-auth-httplib2>=0.1.0
